+++ src/polyline-geography.h
@@ -14,6 +14,27 @@ public:
   PolylineGeography(std::vector<std::unique_ptr<S2Polyline>> polylines):
     polylines(std::move(polylines)) {}
 
+  Geography::Type GeographyType() {
+    return Geography::Type::GEOGRAPHY_POLYLINE;
+  }
+
+  bool FindValidationError(S2Error* error) {
+    bool result;
+    error->Clear();
+    for (size_t i = 0; i < this->polylines.size(); i++) {
+      result = this->polylines[i]->FindValidationError(error);
+      if (result) {
+        return result;
+      }
+    }
+
+    return false;
+  }
+
+  const std::vector<std::unique_ptr<S2Polyline>>* Polyline() {
+    return &(this->polylines);
+  }
+
   bool IsCollection() {
     return this->polylines.size() > 1;
   }