+++ src/s2-accessors.cpp
@@ -18,6 +18,38 @@ LogicalVector cpp_s2_is_collection(List geog) {
 }
 
 // [[Rcpp::export]]
+LogicalVector cpp_s2_is_valid(List geog) {
+  class Op: public UnaryGeographyOperator<LogicalVector, int> {
+    int processFeature(XPtr<Geography> feature, R_xlen_t i) {
+      return !(feature->FindValidationError(&(this->error)));
+    }
+
+    S2Error error;
+  };
+
+  Op op;
+  return op.processVector(geog);
+}
+
+// [[Rcpp::export]]
+CharacterVector cpp_s2_is_valid_reason(List geog) {
+  class Op: public UnaryGeographyOperator<CharacterVector, String> {
+    String processFeature(XPtr<Geography> feature, R_xlen_t i) {
+      if (feature->FindValidationError(&(this->error))) {
+        return this->error.text();
+      } else {
+        return NA_STRING;
+      }
+    }
+
+    S2Error error;
+  };
+
+  Op op;
+  return op.processVector(geog);
+}
+
+// [[Rcpp::export]]
 IntegerVector cpp_s2_dimension(List geog) {
   class Op: public UnaryGeographyOperator<IntegerVector, int> {
     int processFeature(XPtr<Geography> feature, R_xlen_t i) {