+++ src/s2-constructors-formatters.cpp
@@ -72,5 +72,11 @@ List cpp_s2_make_polygon(NumericVector x, NumericVector y,
     reader.iterateFeature();
   }
 
+  if (writer.problemId.size() > 0) {
+    Environment s2NS = Environment::namespace_env("s2");
+    Function stopProblems = s2NS["stop_problems_create"];
+    stopProblems(writer.problemId, writer.problems);
+  }
+
   return writer.output;
 }