+++ src/s2-transformers.cpp
@@ -89,7 +89,7 @@ std::unique_ptr<Geography> doBooleanOperation(S2ShapeIndex* index1, S2ShapeIndex
     options
   );
 
-  // check for errors
+  // build and check for errors
   S2Error error;
   if (!booleanOp.Build(*index1, *index2, &error)) {
     stop(error.text());
@@ -145,7 +145,7 @@ std::unique_ptr<Geography> rebuildGeography(S2ShapeIndex* index,
   // build the output
   S2Error error;
   if (!builder.Build(&error)) {
-    stop(error.text());
+    throw GeographyOperatorException(error.text());
   }
 
   // construct output
@@ -208,7 +208,7 @@ List cpp_s2_sym_difference(List geog1, List geog2, List s2options) {
 }
 
 // [[Rcpp::export]]
-List cpp_s2_union_agg(List geog, List s2options, bool naRm) {
+List cpp_s2_coverage_union_agg(List geog, List s2options, bool naRm) {
   GeographyOperationOptions options(s2options);
 
   MutableS2ShapeIndex index;
@@ -231,6 +231,61 @@ List cpp_s2_union_agg(List geog, List s2options, bool naRm) {
     &emptyIndex,
     S2BooleanOperation::OpType::UNION,
     options.booleanOperationOptions(),
+    options.layerOptions()
+  );
+
+  return List::create(Rcpp::XPtr<Geography>(geography.release()));
+}
+
+// This approach to aggregation is slow but accurate. There is probably a more efficient way
+// to accumulate geometries and/or re-use the layers vector but thus far I haven't figured
+// out a way to make that work.
+// [[Rcpp::export]]
+List cpp_s2_union_agg(List geog, List s2options, bool naRm) {
+  GeographyOperationOptions options(s2options);
+  GeographyOperationOptions::LayerOptions layerOptions = options.layerOptions();
+  S2BooleanOperation::Options unionOptions = options.booleanOperationOptions();
+  S2Builder::Options buillderOptions = options.builderOptions();
+
+  // using smart pointers here so that we can use swap() to
+  // use replace accumulatedIndex with index after each union
+  std::unique_ptr<MutableS2ShapeIndex> index = absl::make_unique<MutableS2ShapeIndex>();
+  std::unique_ptr<MutableS2ShapeIndex> accumulatedIndex = absl::make_unique<MutableS2ShapeIndex>();
+
+  SEXP item;
+  for (R_xlen_t i = 0; i < geog.size(); i++) {
+    item = geog[i];
+    if (item == R_NilValue && !naRm) {
+      return List::create(R_NilValue);
+    }
+
+    if (item != R_NilValue) {
+      Rcpp::XPtr<Geography> feature(item);
+
+      index->Clear();
+      s2builderutil::LayerVector layers(3);
+      layers[0] = absl::make_unique<s2builderutil::IndexedS2PointVectorLayer>(index.get(), layerOptions.pointLayerOptions);
+      layers[1] = absl::make_unique<s2builderutil::IndexedS2PolylineVectorLayer>(index.get(), layerOptions.polylineLayerOptions);
+      layers[2] = absl::make_unique<s2builderutil::IndexedS2PolygonLayer>(index.get(), layerOptions.polygonLayerOptions);
+
+      S2BooleanOperation booleanOp(
+        S2BooleanOperation::OpType::UNION,
+        s2builderutil::NormalizeClosedSet(std::move(layers)),
+        unionOptions
+      );
+
+      S2Error error;
+      if (!booleanOp.Build(*accumulatedIndex, *(feature->ShapeIndex()), &error)) {
+        stop(error.text());
+      }
+
+      accumulatedIndex.swap(index);
+    }
+  }
+
+  std::unique_ptr<Geography> geography = rebuildGeography(
+    accumulatedIndex.get(),
+    options.builderOptions(),
     options.layerOptions()
   );
 
@@ -265,6 +320,33 @@ List cpp_s2_centroid_agg(List geog, bool naRm) {
   }
 
   return output;
+}
+
+// [[Rcpp::export]]
+List cpp_s2_rebuild_agg(List geog, List s2options, bool naRm) {
+  GeographyOperationOptions options(s2options);
+
+  MutableS2ShapeIndex index;
+  SEXP item;
+  for (R_xlen_t i = 0; i < geog.size(); i++) {
+    item = geog[i];
+    if (item == R_NilValue && !naRm) {
+      return List::create(R_NilValue);
+    }
+
+    if (item != R_NilValue) {
+      Rcpp::XPtr<Geography> feature(item);
+      feature->BuildShapeIndex(&index);
+    }
+  }
+
+  std::unique_ptr<Geography> geography = rebuildGeography(
+    &index,
+    options.builderOptions(),
+    options.layerOptions()
+  );
+
+  return List::create(Rcpp::XPtr<Geography>(geography.release()));
 }
 
 std::vector<S2Point> findClosestPoints(S2ShapeIndex* index1, S2ShapeIndex* index2) {
@@ -402,6 +484,114 @@ List cpp_s2_rebuild(List geog, List s2options) {
 
   private:
     S2Builder::Options options;
+    GeographyOperationOptions::LayerOptions layerOptions;
+  };
+
+  Op op(s2options);
+  return op.processVector(geog);
+}
+
+// [[Rcpp::export]]
+List cpp_s2_unary_union(List geog, List s2options) {
+  class Op: public UnaryGeographyOperator<List, SEXP> {
+  public:
+    Op(List s2options) {
+      GeographyOperationOptions options(s2options);
+      this->options = options.booleanOperationOptions();
+      this->layerOptions = options.layerOptions();
+    }
+
+    SEXP processFeature(XPtr<Geography> feature, R_xlen_t i) {
+      // complex union only needed when a polygon is involved
+      bool simpleUnionOK = feature->IsEmpty() || 
+        (feature->Dimension() < 2);
+
+      // valid polygons that are not part of a collection can also use a
+      // simple union (common)
+      if (feature->GeographyType() == Geography::Type::GEOGRAPHY_POLYGON) {
+        S2Error validationError;
+        if(!(feature->Polygon()->FindValidationError(&validationError))) {
+          simpleUnionOK = true;
+        }
+      }
+
+      if (simpleUnionOK) {
+        MutableS2ShapeIndex emptyIndex;
+
+        std::unique_ptr<Geography> ptr = doBooleanOperation(
+          feature->ShapeIndex(),
+          &emptyIndex,
+          S2BooleanOperation::OpType::UNION,
+          this->options,
+          this->layerOptions
+        );
+
+        return XPtr<Geography>(ptr.release());
+      } else if (feature->GeographyType() == Geography::Type::GEOGRAPHY_POLYGON) {
+        // If we've made it here we have an invalid polygon on our hands. A geography with
+        // invalid loops won't work with the S2BooleanOperation we will use to accumulate
+        // (i.e., union) valid polygons, so we need to rebuild each loop as its own polygon,
+        // splitting crossed edges along the way.
+        const S2Polygon* originalPoly = feature->Polygon();
+        
+        // Not exposing these options as an argument (except snap function)
+        // because a particular combiation of them is required for this to work
+        S2Builder::Options builderOptions;
+        builderOptions.set_split_crossing_edges(true);
+        builderOptions.set_snap_function(this->options.snap_function());
+        s2builderutil::S2PolygonLayer::Options layerOptions;
+        layerOptions.set_edge_type(S2Builder::EdgeType::UNDIRECTED);
+        layerOptions.set_validate(false);
+
+        // Rebuild all loops as polygons using the S2Builder()
+        std::vector<std::unique_ptr<S2Polygon>> loops;
+        for (int i = 0; i < originalPoly->num_loops(); i++) {
+          std::unique_ptr<S2Polygon> loop = absl::make_unique<S2Polygon>();
+          S2Builder builder(builderOptions);
+          builder.StartLayer(absl::make_unique<s2builderutil::S2PolygonLayer>(loop.get()));
+          builder.AddShape(S2Loop::Shape(originalPoly->loop(i)));
+          S2Error error;
+          if (!builder.Build(&error)) {
+            throw GeographyOperatorException(error.text());
+          }
+
+          // Check if the builder created a polygon whose boundary contained more than
+          // half the earth (and invert it if so)
+          if (loop->GetArea() > (2 * PI)) {
+            loop->Invert();
+          }
+
+          loops.push_back(std::move(loop));
+        }
+
+        // Accumulate the union of outer loops (but difference of inner loops)
+        std::unique_ptr<S2Polygon> accumulatedPolygon = absl::make_unique<S2Polygon>();
+        for (int i = 0; i < originalPoly->num_loops(); i++) {
+          std::unique_ptr<S2Polygon> polygonResult = absl::make_unique<S2Polygon>();
+
+          // Use original nesting to suggest if this loop should be unioned or diffed.
+          // For valid polygons loops are arranged such that the biggest loop is on the outside
+          // followed by holes such that the below strategy should work (since we are
+          // just iterating along the original loop structure)
+          if ((originalPoly->loop(i)->depth() % 2) == 0) {
+            polygonResult->InitToUnion(accumulatedPolygon.get(), loops[i].get());
+          } else {
+            polygonResult->InitToDifference(accumulatedPolygon.get(), loops[i].get());
+          }
+
+          accumulatedPolygon.swap(polygonResult);
+        }
+
+        return XPtr<Geography>(new PolygonGeography(std::move(accumulatedPolygon)));
+      } else {
+        // This is a less common case (mixed dimension output that includes a polygon).
+        // In the absence of a clean solution, saving this battle for another day.
+        throw GeographyOperatorException("Unary union for collections is not implemented");
+      }
+    }
+
+  private:
+    S2BooleanOperation::Options options;
     GeographyOperationOptions::LayerOptions layerOptions;
   };