+++ NEWS.md
@@ -1,5 +1,25 @@
 # s2 (development version)
 
+* Added `s2_is_valid()` and `s2_is_valid_detail()` to help find invalid
+  spherical geometries when importing data into S2 (#100).
+* Improved error messages when importing and processing data such that
+  errors can be debugged more readily (#100, #98).
+* The unary version of `s2_union()` can now handle MULTIPOLYGON
+  geometries with overlapping rings in addition to other invalid
+  polygons. `s2_union()` can now sanitize
+  almost any input to be valid spherical geometry with
+  minimal modification (#100, #99).
+* Renamed the existing implementation of `s2_union_agg()` to
+  `s2_coverage_union_agg()` to make clear that the function only
+  works when the individual geometries do not have overlapping
+  interiors. `s2_union_agg()` was replaced with a
+  true aggregate union that can handle unions of most geometries
+  (#100, #97).
+* Added `s2_rebuild_agg()` to match `s2_union_agg()`. Like
+  `s2_rebuild()`, `s2_rebuild_agg()` collects the edges in the input
+  and builds them into a feature, optionally snapping or simplifying
+  vertices in the process (#100).
+
 # s2 1.0.4
 
 * Fixed errors that resulted from compilation on clang 12.2 (#88, #89).